#121 Add ability to edit role and delete user by admin

+++ backend/src/api/routes/admin-route.ts
@@ -6,9 +6,15 @@ import { createTagsController } from '../controllers/tags.controller';
 import { registerUserController } from '../controllers/user-managment.controller';
 
 import { deleteTag, getTags } from '~/services/tag.service';
-import { getCommonUserList, deleteUser } from '~/services/user.service';
+import {
+  getCommonUserList,
+  deleteUser,
+  changeUserRole,
+} from '~/services/user.service';
 
 import { Headers } from '~/common/enums/headers';
+import { validatePermissions } from '../middlewares/validation-middleware';
+import { RoleType } from '~/common/enums/role-type';
 
 const router: Router = Router();
 
@@ -26,6 +32,7 @@ router.delete(
 
 router.get(
   '/users',
+  validatePermissions([RoleType.ADMIN]),
   run((req) => getCommonUserList(req.companyId)),
 );
 
@@ -43,7 +50,13 @@ router.post(
 
 router.delete(
   '/users/:id',
+  validatePermissions([RoleType.ADMIN]),
   run((req) => deleteUser(req.params.id)),
 );
 
+router.put(
+  '/users/:id',
+  validatePermissions([RoleType.ADMIN]),
+  run((req) => changeUserRole(req.params.id, req.body)),
+);
 export default router;
